Add FromIteratorSpec
