+++ creusot/tests/should_succeed/100doors.mlcfg
@@ -552,19 +552,19 @@ module CreusotContracts_Resolve_Resolve_Resolve
     ensures { result = resolve self }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -719,7 +719,7 @@ module C100doors_F
     type t = Alloc_Vec_Vec_Type.t_vec bool (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = bool,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
