+++ creusot/tests/should_succeed/index_range.mlcfg
@@ -214,19 +214,19 @@ module CreusotContracts_Resolve_Resolve_Resolve
     ensures { result = resolve self }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -294,7 +294,7 @@ module IndexRange_CreateArr
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve1.resolve
@@ -1120,7 +1120,7 @@ module IndexRange_TestRange
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -1998,7 +1998,7 @@ module IndexRange_TestRangeTo
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -2643,7 +2643,7 @@ module IndexRange_TestRangeFrom
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -3294,7 +3294,7 @@ module IndexRange_TestRangeFull
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -3922,7 +3922,7 @@ module IndexRange_TestRangeToInclusive
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
