+++ creusot/tests/should_succeed/iterators/08_collect_extend.mlcfg
@@ -774,19 +774,19 @@ module Alloc_Vec_Impl0_New
     ensures { Seq.length (ShallowModel0.shallow_model result) = 0 }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -878,7 +878,7 @@ module C08CollectExtend_Collect
     type self = Item0.item
   clone CreusotContracts_Resolve_Resolve_Resolve as Resolve4 with
     type self = Core_Option_Option_Type.t_option Item0.item
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve3 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve3 with
     type t = Item0.item,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve5.resolve
@@ -1406,7 +1406,7 @@ module C08CollectExtend_ExtendIndex
     function ShallowModel0.shallow_model = ShallowModel1.shallow_model,
     function ShallowModel1.shallow_model = ShallowModel0.shallow_model,
     predicate Invariant0.invariant' = Invariant0.invariant'
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = uint32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -1523,7 +1523,7 @@ module C08CollectExtend_CollectExample
     type t = uint32,
     type a = Alloc_Alloc_Global_Type.t_global,
     axiom .
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = uint32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
