+++ creusot/tests/should_succeed/knapsack_full.mlcfg
@@ -918,19 +918,19 @@ module CreusotContracts_Resolve_Impl1_Resolve
     ensures { result = resolve self }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -1125,7 +1125,7 @@ module KnapsackFull_Knapsack01Dyn
     type t = usize,
     type a = Alloc_Alloc_Global_Type.t_global,
     axiom .
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve6 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve6 with
     type t = usize,
     function ShallowModel0.shallow_model = ShallowModel4.shallow_model,
     predicate Resolve0.resolve = Resolve7.resolve
@@ -1190,13 +1190,13 @@ module KnapsackFull_Knapsack01Dyn
     type t = KnapsackFull_Item_Type.t_item name
   clone CreusotContracts_Std1_Slice_Impl5_InBounds as InBounds0 with
     type t = KnapsackFull_Item_Type.t_item name
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve5 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve5 with
     type t = Alloc_Vec_Vec_Type.t_vec usize (Alloc_Alloc_Global_Type.t_global),
     function ShallowModel0.shallow_model = ShallowModel3.shallow_model,
     predicate Resolve0.resolve = Resolve6.resolve
   clone CreusotContracts_Resolve_Resolve_Resolve as Resolve0 with
     type self = KnapsackFull_Item_Type.t_item name
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve4 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve4 with
     type t = KnapsackFull_Item_Type.t_item name,
     function ShallowModel0.shallow_model = ShallowModel1.shallow_model,
     predicate Resolve0.resolve = Resolve0.resolve
