+++ creusot/tests/should_succeed/syntax/12_ghost_code.mlcfg
@@ -145,19 +145,19 @@ module CreusotContracts_Resolve_Resolve_Resolve
     ensures { result = resolve self }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -206,7 +206,7 @@ module C12GhostCode_GhostVec
     type t = uint32,
     type a = Alloc_Alloc_Global_Type.t_global,
     axiom .
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = uint32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve1.resolve
@@ -502,7 +502,7 @@ module C12GhostCode_GhostCheck
     type t = Alloc_Vec_Vec_Type.t_vec int32 (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = int32,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve1.resolve
