+++ creusot/tests/should_succeed/vector/06_knights_tour.mlcfg
@@ -431,19 +431,19 @@ module CreusotContracts_Resolve_Resolve_Resolve
     ensures { result = resolve self }
     
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Stub
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Stub
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve_Interface
+module CreusotContracts_Std1_Vec_Impl6_Resolve_Interface
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
   predicate resolve (self : Alloc_Vec_Vec_Type.t_vec t (Alloc_Alloc_Global_Type.t_global))
 end
-module CreusotContracts_Std1_Vec_Impl5_Resolve
+module CreusotContracts_Std1_Vec_Impl6_Resolve
   type t
   use Alloc_Alloc_Global_Type as Alloc_Alloc_Global_Type
   use Alloc_Vec_Vec_Type as Alloc_Vec_Vec_Type
@@ -511,7 +511,7 @@ module C06KnightsTour_Impl1_New
   use seq.Seq
   clone CreusotContracts_Resolve_Impl2_Resolve as Resolve2 with
     type t = usize
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = usize,
     function ShallowModel0.shallow_model = ShallowModel1.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
@@ -522,7 +522,7 @@ module C06KnightsTour_Impl1_New
     type t = Alloc_Vec_Vec_Type.t_vec (Alloc_Vec_Vec_Type.t_vec usize (Alloc_Alloc_Global_Type.t_global)) (Alloc_Alloc_Global_Type.t_global),
     type ShallowModelTy0.shallowModelTy = ShallowModelTy0.shallowModelTy,
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = Alloc_Vec_Vec_Type.t_vec usize (Alloc_Alloc_Global_Type.t_global),
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve1.resolve
@@ -1243,7 +1243,7 @@ module C06KnightsTour_Impl1_CountDegree
     predicate InBounds0.in_bounds = InBounds1.in_bounds,
     predicate HasValue0.has_value = HasValue0.has_value,
     type Output0.output = Output0.output
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = (isize, isize),
     function ShallowModel0.shallow_model = ShallowModel2.shallow_model,
     predicate Resolve0.resolve = Resolve1.resolve
@@ -2032,7 +2032,7 @@ module C06KnightsTour_KnightsTour
     type t = Alloc_Vec_Vec_Type.t_vec usize (Alloc_Alloc_Global_Type.t_global),
     type a = Alloc_Alloc_Global_Type.t_global,
     axiom .
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve1 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve1 with
     type t = (usize, C06KnightsTour_Point_Type.t_point),
     function ShallowModel0.shallow_model = ShallowModel0.shallow_model,
     predicate Resolve0.resolve = Resolve3.resolve
@@ -2062,7 +2062,7 @@ module C06KnightsTour_KnightsTour
     predicate InBounds0.in_bounds = InBounds1.in_bounds,
     predicate HasValue0.has_value = HasValue0.has_value,
     type Output0.output = Output0.output
-  clone CreusotContracts_Std1_Vec_Impl5_Resolve as Resolve0 with
+  clone CreusotContracts_Std1_Vec_Impl6_Resolve as Resolve0 with
     type t = (isize, isize),
     function ShallowModel0.shallow_model = ShallowModel3.shallow_model,
     predicate Resolve0.resolve = Resolve2.resolve
